move to shape_inflate.{hpp, cpp}
